+++ src/structures/paging/page.rs
@@ -206,7 +206,7 @@ impl<S: PageSize> fmt::Debug for Page<S> {
 impl<S: PageSize> Add<u64> for Page<S> {
     type Output = Self;
     fn add(self, rhs: u64) -> Self::Output {
-        Page::containing_address(self.start_address() + rhs * u64::from(S::SIZE))
+        Page::containing_address(self.start_address() + rhs * S::SIZE)
     }
 }
 
@@ -219,7 +219,7 @@ impl<S: PageSize> AddAssign<u64> for Page<S> {
 impl<S: PageSize> Sub<u64> for Page<S> {
     type Output = Self;
     fn sub(self, rhs: u64) -> Self::Output {
-        Page::containing_address(self.start_address() - rhs * u64::from(S::SIZE))
+        Page::containing_address(self.start_address() - rhs * S::SIZE)
     }
 }
